fix(RBACProvider): Move to components package
