+++ packages/components/src/RBACProvider/index.ts
@@ -1,0 +1,1 @@
+export { default as RBACProvider } from './RBACProvider';
