+++ packages/utils/src/RBACHook/index.ts
@@ -1,3 +1,2 @@
-export { default as RBACProvider } from './RBACProvider';
 export { default } from './RBACHook';
 export * from './RBACHook';
